DEPgraphql/member-service/requirements.txt
fastapi
uvicorn
sqlalchemy
strawberry-graphql
requests